pyproject.toml
[project]
name = "agent-mark-landing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
    "uvicorn>=0.34.0",
]
